+++ doc/whats-new.rst
@@ -43,6 +43,8 @@ Bug fixes
 - Aggregation functions now correctly skip ``NaN`` for data for ``complex128``
   dtype (:issue:`554`).
 - Fixed indexing 0d arrays with unicode dtype (:issue:`568`).
+- :py:meth:`~xray.DataArray.name` and Dataset keys must be a string or None to
+  be written to netCDF (:issue:`533`).
 
 v0.6.0 (21 August 2015)
 -----------------------